Cast to int.

+++ app/Http/Controllers/Api/V1/KeywordController.php
@@ -76,7 +76,7 @@ class KeywordController extends Controller
             $input = $request->all();
             $jwtUser = array_key_exists('jwt_user', $input) ? $input['jwt_user'] : [];
 
-            $perPage = request('perPage', Config::get('constants.per_page'));
+            $perPage = (int) request('perPage', Config::get('constants.per_page'));
             $keywords = Keyword::where('enabled', 1)
                 ->paginate(function ($total) use ($perPage) {
                     if($perPage === -1) {
