+++ app/Http/Controllers/Api/V1/ProgrammingLanguageController.php
@@ -48,7 +48,7 @@ class ProgrammingLanguageController extends Controller
      */
     public function index(Request $request): JsonResponse
     {
-        $perPage = request('per_page', Config::get('constants.per_page'));
+        $perPage = (int) request('per_page', Config::get('constants.per_page'));
         try {
             $input = $request->all();
             $jwtUser = array_key_exists('jwt_user', $input) ? $input['jwt_user'] : [];
